renamed package directive
