+++ service/cloudconfigv2/cloud_config.go
@@ -1,4 +1,4 @@
-package cloudconfigv1
+package cloudconfigv2
 
 import (
 	"github.com/giantswarm/certificatetpr"
