+++ service/cloudconfigv2/cloudconfigtest/cloudconfigtest.go
@@ -3,15 +3,15 @@ package cloudconfigtest
 import (
 	"github.com/giantswarm/micrologger/microloggertest"
 
-	"github.com/giantswarm/kvm-operator/service/cloudconfigv1"
+	"github.com/giantswarm/kvm-operator/service/cloudconfigv2"
 )
 
-func New() *cloudconfigv1.CloudConfig {
-	c := cloudconfigv1.DefaultConfig()
+func New() *cloudconfigv2.CloudConfig {
+	c := cloudconfigv2.DefaultConfig()
 
 	c.Logger = microloggertest.New()
 
-	newCloudConfig, err := cloudconfigv1.New(c)
+	newCloudConfig, err := cloudconfigv2.New(c)
 	if err != nil {
 		panic(err)
 	}
