+++ service/cloudconfigv2/v_1_0_0_master_template.go
@@ -1,4 +1,4 @@
-package cloudconfigv1
+package cloudconfigv2
 
 import (
 	"github.com/giantswarm/certificatetpr"
